+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Newsletter link on socials bottom sheet.
 
+### Fixed
+
+- Properly reset animation on Home screen once refresh is done.
+
 ## [2025.4.1] - 2025-04-15
 
 ### Added